refactor: earningRate 정수로 처리
- 부동소숫점 방식보다, 정수로 처리한 후 100으로 나누는 것이 좀 더 정확할 것 같아서 판단 (소수점 버림 연산)
- BettingResultTest에서 earningRate를 가져와서 계산하기 보다는, 그 자체의 계산에 대해서 테스트를 수행
- BettingResultTest의 중복된 객체 생성에 대해서 @BeforeEach 사용
- 공백 수정

+++ src/main/java/service/BettingCalculateService.java
@@ -26,7 +26,7 @@ public class BettingCalculateService {
         for (Player player : players) {
             Money bettingMoney = player.getBettingMoney();
             BettingResult judge = BettingResult.judge(player, dealer);
-            double earningsRate = judge.getEarningRate();
+            int earningsRate = judge.getEarningRate();
             playersProfit.add(new Profit(player.getName(), bettingMoney, earningsRate));
         }
         return playersProfit;
